refactor(server): split routing tests and muc delivery into focused modules
Code-motion-only round-4 follow-up: interpret/tests/routing.rs
(1,290 lines) becomes six routing_* modules by disposition
(route_to_connection selection, full-JID fallback, detached delivery,
local-account offline/bounce, negative priority, fanout pass);
delivery/muc.rs becomes muc/{proxy,reserved,types}.rs behind a thin
facade. Function inventory unchanged; largest new module is 348 lines.
